Don't purge the resource status cache upon docker wrapped commands
These are not necessarily state-changing action.

This patch is necessary to be able to trigger the resource restart after
a manual docker container stop.

+++ lib/svc.py
@@ -3549,7 +3549,8 @@ class Svc(object):
             # purge the resource status file cache, so that we don't take
             # decision on outdated information
             #
-            if not options.dry_run and action != "resource_monitor":
+            if not options.dry_run and action != "resource_monitor" and \
+               not action.startswith("docker"):
                 self.log.debug("purge all resource status file caches")
                 self.purge_status_last()
